# docs/api/openapi-fase1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HERO Backend API",
    "version": "0.10.0"
  },
  "paths": {
...
    "/api/v1/documents/{document_id}/pdf": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Buka atau unduh berkas PDF asli",
        "description": "Mengembalikan berkas PDF asli dari penyimpanan.\nMencatat audit OPEN_PDF atau DOWNLOAD_PDF.",
        "operationId": "get_document_pdf_api_v1_documents__document_id__pdf_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          },
          {
            "name": "download",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Jika true, set Content-Disposition ke attachment",
              "default": false,
              "title": "Download"
            },
            "description": "Jika true, set Content-Disposition ke attachment"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "Dokumen non-publik memerlukan autentikasi aktif"
          },
          "404": {
            "description": "Dokumen atau berkas fisik PDF tidak ditemukan"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/articles": {
      "post": {
        "tags": [
          "Internal"
        ],
        "summary": "[Internal] Bulk insert chunk pasal dari ML pipeline",
        "description": "Bulk insert pasal dari pipeline ML.",
        "operationId": "bulk_insert_articles_api_v1_internal_articles_post",
        "parameters": [
          {
            "name": "X-Internal-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkArticleIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkArticleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/extraction/claim": {
      "post": {
        "tags": [
          "Internal"
        ],
        "summary": "[Internal] Klaim antrean dokumen untuk diekstraksi oleh Data/ML",
        "description": "Mengambil dan mengunci dokumen dalam antrean ekstraksi (SELECT FOR UPDATE SKIP LOCKED).\nMencakup status 'diterima' dan status 'diproses' yang telah kedaluwarsa (claim timeout).",
        "operationId": "claim_extraction_batch_api_v1_internal_extraction_claim_post",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Jumlah dokumen yang ingin diklaim",
              "default": 10,
              "title": "Limit"
            },
            "description": "Jumlah dokumen yang ingin diklaim"
          },
          {
            "name": "X-Internal-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/documents/{document_id}/pdf": {
      "get": {
        "tags": [
          "Internal"
        ],
        "summary": "[Internal] Unduh berkas PDF untuk ekstraksi",
        "description": "Mengembalikan berkas PDF asli untuk kebutuhan worker ekstraksi Data/ML tanpa audit per unduhan.",
        "operationId": "get_internal_document_pdf_api_v1_internal_documents__document_id__pdf_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          },
          {
            "name": "X-Internal-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/documents/{document_id}/extraction": {
      "patch": {
        "tags": [
          "Internal"
        ],
        "summary": "[Internal] Kirim hasil ekstraksi metadata dan teks dari Data/ML",
        "description": "Menerima hasil ekstraksi Data/ML (US-20, US-20a, US-26).\nMendukung penanganan error ekstraksi, proteksi koreksi manual, dan penempatan otomatis.",
        "operationId": "patch_extraction_result_api_v1_internal_documents__document_id__extraction_patch",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Paksa update meskipun status dokumen bukan 'diproses'",
              "default": false,
              "title": "Force"
            },
            "description": "Paksa update meskipun status dokumen bukan 'diproses'"
          },
          {
            "name": "X-Internal-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractionResultIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/extraction/requeue/{document_id}": {
      "post": {
        "tags": [
          "Internal"
        ],
        "summary": "[Internal] Mengembalikan dokumen ke antrean ekstraksi",
        "description": "Mengembalikan dokumen ke status 'diterima' dan mereset claimed_at (tidak mereset attempt count).",
        "operationId": "requeue_document_extraction_api_v1_internal_extraction_requeue__document_id__post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          },
          {
            "name": "X-Internal-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/scans/claim": {
      "post": {
        "tags": [
          "Internal"
        ],
        "summary": "[Internal Data/ML] Mengklaim antrean sesi pemindaian mode push",
        "description": "[Langkah 7 - Opsi B Data/ML] Mengambil sesi pemindaian antrean dengan mode='push'.",
        "operationId": "claim_push_scans_endpoint_api_v1_internal_scans_claim_post",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5,
              "minimum": 1,
              "description": "Jumlah sesi yang ingin diklaim (1-5)",
              "default": 1,
              "title": "Limit"
            },
            "description": "Jumlah sesi yang ingin diklaim (1-5)"
          },
          {
            "name": "X-Internal-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/scans/{scan_id}/candidates": {
      "post": {
        "tags": [
          "Internal"
        ],
        "summary": "[Internal Data/ML] Mengirim batch kandidat PDF hasil pemindaian",
        "description": "[Langkah 7 - Opsi B Data/ML] Menerima kandidat PDF yang ditemukan oleh crawler eksternal.\nSaat done=true, sistem otomatis menjalankan perbandingan dengan basis pengetahuan.",
        "operationId": "push_scan_candidates_endpoint_api_v1_internal_scans__scan_id__candidates_post",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Scan Id"
            }
          },
          {
            "name": "X-Internal-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InternalCandidatesBatchIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkArticleIn": {
        "properties": {
          "articles": {
            "items": {
              "$ref": "#/components/schemas/ArticleChunkIn"
            },
            "type": "array",
            "minItems": 1,
            "title": "Articles",
            "description": "Daftar chunk pasal yang akan dimasukkan ke database (min. 1 item)"
          }
        },
        "type": "object",
        "required": [
          "articles"
        ],
        "title": "BulkArticleIn",
        "description": "Payload untuk bulk insert: list chunk pasal dari satu batch ML pipeline."
      },
      "BulkArticleResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Status operasi",
            "default": "ok"
          },
          "inserted_count": {
            "type": "integer",
            "title": "Inserted Count",
            "description": "Jumlah pasal yang berhasil disimpan"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Pesan ringkasan"
          }
        },
        "type": "object",
        "required": [
          "inserted_count",
          "message"
        ],
        "title": "BulkArticleResponse",
        "description": "Response setelah bulk insert berhasil."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}